Refactor i18n init to run before presentation layer
