+++ src/presentation/widget/WidgetApp.jsx
@@ -1,14 +1,13 @@
-import { useConfig, useDataQuery } from "@dhis2/app-runtime";
-import i18n from "@dhis2/d2-i18n";
+import { useConfig } from "@dhis2/app-runtime";
 import { MuiThemeProvider } from "@material-ui/core/styles";
 import { createGenerateClassName, StylesProvider } from "@material-ui/styles";
 import axiosRetry from "axios-retry";
 import { init } from "d2";
 import { LoadingProvider, SnackbarProvider } from "d2-ui-components";
-import _ from "lodash";
 import OldMuiThemeProvider from "material-ui/styles/MuiThemeProvider";
 import React, { useEffect, useState } from "react";
 import { Instance } from "../../domain/instance/entities/Instance";
+import i18n from "../../locales";
 import { MigrationsRunner } from "../../migrations";
 import { D2Api } from "../../types/d2-api";
 import { AppContext } from "../common/contexts/AppContext";
@@ -24,26 +23,10 @@ const generateClassName = createGenerateClassName({
     productionPrefix: "c",
 });
 
-const isLangRTL = code => {
-    const langs = ["ar", "fa", "ur"];
-    const prefixed = langs.map(c => `${c}-`);
-    return _(langs).includes(code) || prefixed.filter(c => code && code.startsWith(c)).length > 0;
-};
-
-const configI18n = ({ keyUiLocale: uiLocale }) => {
-    i18n.changeLanguage(uiLocale);
-    document.documentElement.setAttribute("dir", isLangRTL(uiLocale) ? "rtl" : "ltr");
-};
-
-const query = {
-    userSettings: { resource: "/userSettings" },
-};
-
 const App = () => {
     const { baseUrl } = useConfig();
     const [appContext, setAppContext] = useState(null);
     const [migrationsState, setMigrationsState] = useState({ type: "checking" });
-    const { error, data } = useDataQuery(query);
 
     useEffect(() => {
         const run = async () => {
@@ -61,16 +44,13 @@ const App = () => {
             const compositionRoot = new CompositionRoot(instance, d2, encryptionKey);
             setAppContext({ d2, api, compositionRoot });
 
-            configI18n(data.userSettings);
             runMigrations(api).then(setMigrationsState);
         };
 
-        if (data) run();
-    }, [data, baseUrl]);
+        run();
+    }, [baseUrl]);
 
-    if (error) {
-        return <p>{i18n.t("Unknown DHIS2 error")}</p>;
-    } else if (migrationsState.type === "pending") {
+    if (migrationsState.type === "pending") {
         return (
             <p>{i18n.t("Widget cannot be used until an administrator opens the application")}</p>
         );
